% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{buscar_marmitas_veganas}
\alias{buscar_marmitas_veganas}
\title{Exibir as marmitas vegenas disponiveis no beleaf}
\usage{
buscar_marmitas_veganas()
}
\value{
retorna um tibble com 38 linhas e 6 colunas
}
\description{
Exibir as marmitas vegenas disponiveis no beleaf
}
\examples{
# exibir cardapio
buscar_marmitas_veganas()
}
