% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{marmita_vegana}
\alias{marmita_vegana}
\title{Cardapio de comida vegana}
\format{
Um tibble com 39 linhas e 6 colunas:
\describe{
\item{item}{Enumeracao de cada um dos pratos}
\item{nome}{Nome do prato}
\item{descricao}{Ingredientes utilizados na preparacao do prato}
\item{preco}{Preco em reais dos pratos}
\item{tipo_refeicao}{Especificacao da refeicao: prato, sopa e doce}
\item{url}{link para acesso do prato na web}
}
}
\source{
\url{"https://www.beleaf.com.br/cardapio-marmitas-veganas"}
}
\usage{
marmita_vegana
}
\description{
Caradapio de comida vegana disponivel no beleaf
}
\keyword{datasets}
