% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{sortear_marmita}
\alias{sortear_marmita}
\title{Sortear aleatoriamente uma marmita disponivel no beleaf a partir do(s) tipo(s) de refeicao(oes) e de um preco maximo}
\usage{
sortear_marmita(tipo = c("Sopa", "Prato", "Doce"), preco_maximo = 50)
}
\arguments{
\item{tipo}{Um vetor de caracteres informando a refeicao: As opcoes disponiveis sao: "Sopa", "Prato", "Doce". O default e c("Sopa", "Prato", "Doce")}

\item{preco_maximo}{Um inteiro maior que zero}
}
\value{
Um vetor de caracteres
}
\description{
Sortear aleatoriamente uma marmita disponivel no beleaf a partir do(s) tipo(s) de refeicao(oes) e de um preco maximo
}
\examples{
# saida default da funcao
sortear_marmita()

# opcoes de sopa
sortear_marmita(tipo = "Sopa")

# opcoes de Doce
sortear_marmita(tipo = "Doce")

# refeicoes com preco de no maximo R$ 20,00
sortear_marmita(preco_maximo = 20)
}
